Finalizado Projeto de Filmes

+++ FrontEnd-3S/ReactJS/filmes/src/pages/login/login.jsx
@@ -1,33 +1,79 @@
 import Logo from "../../assets/img/logo.svg";
 import "./Login.css";
 import Botao from "../../components/botao/Botao.jsx";
-import { useContext } from "react";
+import { useContext, useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { UsuarioContext } from "../../context/UsuarioContext";
+import { Alerta } from "../../components/alerta/Alerta";
+import api from "../../services/services";
+import { jwtDecode } from "jwt-decode";
 
 const Login = () => {
 
     const { setUsuario } = useContext(UsuarioContext);
-
     const navigate = useNavigate();
 
-    const realizarLogin = (e) => {
+    const [email, setEmail] = useState("");
+    const [senha, setSenha] = useState("");
 
+    const login = async (e) => {
         e.preventDefault();
 
+        if (email.trim() === "" || senha.trim() === "") {
+            Alerta({
+                title: "Login",
+                text: "Preencher todos os campos!",
+                icon: "warning",
+                confirmButtonText: "OK",
+            });
+            return;
+        }
+
         const usuario = {
-            nome: "Marcos"
+            email,
+            senha,
         };
 
-        localStorage.setItem(
-            "usuario",
-            JSON.stringify(usuario)
-        );
+        try {
+            const retornoAPI = await api.post("/Login", usuario);
 
-        setUsuario(usuario);
+            const token = retornoAPI.data.token;
+            const usuarioDecoded = jwtDecode(token);
 
-        navigate("/filmes");
+            setUsuario(usuarioDecoded);
+
+            localStorage.setItem(
+                "token",
+                JSON.stringify(token)
+            );
+
+            setEmail("");
+            setSenha("");
+
+            navigate("/filmes");
+
+        } catch (error) {
+            Alerta({
+                title: "Login",
+                text: "Usuário não encontrado!",
+                icon: "warning",
+                confirmButtonText: "OK",
+            });
+        }
     };
+
+    const verificarLogin = () => {
+        const logado = JSON.parse(localStorage.getItem("usuario"))
+
+        if (logado != undefined || logado != null) {
+            setUsuario(logado);
+            navigate("/generos");
+        }
+    }
+
+    useEffect(() => {
+        verificarLogin();
+    }, []);
 
     return (
         <main className="main_login">
@@ -43,7 +89,7 @@ const Login = () => {
 
                 <form
                     className="form_login"
-                    onSubmit={realizarLogin}
+                    onSubmit={login}
                 >
 
                     <h1>Login</h1>
@@ -60,7 +106,10 @@ const Login = () => {
                                 type="email"
                                 id="email"
                                 name="email"
+                                value={email}
+                                onChange={(e) => setEmail(e.target.value)}
                                 placeholder="Digite seu e-mail"
+                                required
                             />
 
                         </div>
@@ -75,7 +124,10 @@ const Login = () => {
                                 type="password"
                                 id="senha"
                                 name="senha"
+                                value={senha}
+                                onChange={(e) => setSenha(e.target.value)}
                                 placeholder="Digite sua senha"
+                                required
                             />
 
                         </div>
